% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Problem}
\usage{
birthday(n)
}
\arguments{
\item{n}{the number of people in the group}
}
\value{
a probability scalar that 2 or more people sharing a birthday in a group
}
\description{
Birthday Problem
}
\examples{
birthday(25)
}
