% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `who`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{fire damage in thousands of dollars}
  ...
}
}
\source{
<Mendenhall, W. M., & Sincich, T. L. (2016). Statistics for Engineering and the Sciences (6th ed.). Taylor & Francis. https://ou-bookshelf.vitalsource.com/books/9781498728898>
}
\usage{
fire
}
\description{
Data by an insurance company from a study on fires in a suburb ...
}
\keyword{datasets}
