% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootleg Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations of sampling}

\item{x}{original set of elements to draw from}

\item{fun}{function to be estimated}

\item{alpha}{significance level}

\item{cx}{font size}

\item{...}{other}
}
\value{
a histogram of the bootstrap sample statistic wiht the confidence interval and point estimate labelled
}
\description{
Bootleg Function
}
\examples{
myboot2(10000,x=1:10,fun="mean",alpha=0.05,cx=1.5)
}
