% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Exploration of the CLT with Binomial Distribution}
\usage{
mycltb(n = 5, iter = 10000, p = 0.5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations in simulation}

\item{p}{probability}
}
\value{
a histogram of the sample mean with the normal curve overlayed
}
\description{
Exploration of the CLT with Binomial Distribution
}
\examples{
mycltb(n=5,iter=10000,p=0.5)
}
