% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{HYPERGEOMETRIC simulation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{number of simulations run}

\item{N}{total group size (population)}

\item{r}{number of successes in population}

\item{n}{sample size}
}
\value{
A bar plot of the distribution of number of successes over the hypergeometric simulation
}
\description{
HYPERGEOMETRIC simulation
}
\examples{
myhyper(iter=1000,n=5, N=20,r=12)
}
