% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Probability under a normal curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of distribution}

\item{sigma}{standard deviation of distribution}

\item{a}{target X value for probability}
}
\value{
a plot of a normal curve with the probability of X less than or equal to a shaded
}
\description{
Probability under a normal curve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
