% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking problem}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the plane}

\item{gamma}{probability of having more shows than seats (pain)}

\item{p}{probability of a person showing up}
}
\value{
a plot for a discrete function for the problem, a plot for a continuous function, and a list of the number of tickets to sell based on each
}
\description{
Overbooking problem
}
\examples{
ntickets(N=200,gamma=0.8,p=0.95)
}
