% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymle.R
\name{shinymle}
\alias{shinymle}
\title{Maximum Likelihood Estimation for Different Distributions}
\usage{
shinymle()
}
\value{
an app that creates MLE graphs for the selected distribution and sample size n
}
\description{
Maximum Likelihood Estimation for Different Distributions
}
\examples{
shinymle
}
